Library comparison validation
Implemented preliminary validation of comparison of parsed libraries. Client now sends their own parsed library back to the server in response after finishing comparing the received library to their own, along with the number of shared songs that they found. This is received by the server and compared to the server's library, which is checked against the received number from the client.

Reordered control flow to utilize state machines in both client and server. This allows for continuing interactions between them without having to close and reconnect. I'm not the biggest fan of this implementation, as I feel it will get extremely messy as the project gets more complicated and could be more efficiently/cleanly implemented with some sort of asynchronous solution (Twisted uses Deferreds), but this works for the time being.