Preliminary GUI
Created a preliminary implementation of a GUI that will display the available songs to download and allow the user to check ones they would like to download. It uses PyQt5 to display a treeview of a parsed library dictionary, which is internally parsed and stored as a tree.

Changed parse_library_hash function to include whole path of song instead of just name to allow for GUI to construct treeview of parsed library.