Server and client functionality within GUI
Added preliminary server and client functionality within the main GUI. This consists of a number of parts:
* Dedicated server start and connect buttons that both require a parsed library.json file to exist for them to function
* A text entry box for IP if connecting to a server (connect button will not function without IP)
* Parse button will now prompt if a library.json file already exists
* Re-tooled the server and client files to allow for importing and usage within the GUI
* Server and client run within their own threads

Starting the server in a thread gives a strange ValueError about signals within threads that is fully contained within Twisted, so I'm not sure how to go about triaging it. It hasn't given me any issue on my own system, so we'll see how it holds up in the future.